Resolved Beth's issues: 1. Delete the "Pledges to Date" buttons on Pledge Home page and Thank You page
2. Change the first paragraph of the HH Pledge home page to:
Rosh Hashana begins early this year, on September 4th, and CBI is kicking off our High Holiday Appeal.  For the first time we are asking you to make online pledges before the holidays, and for the first time we are asking you to make a spiritual pledge as well as a financial pledge.  The spiritual pledge is your pledge to take on at least one new mitzvah in the coming year.  It can be anything you choose, some options I suggest are to attend a minyan or to donate blood.  This is a pledge to do something for yourself, as well as for the community.  Please consider making both spiritual and financial pledges as we renew ourselves in the New Year.

3. The Spiritual Pledge conformation for one pledge is:
 "Thank you for your pledge of the following mitzvot:"
Should be mitzvah

Also there was an extra dot when I made an "other" pledge:
"Thank you for your pledge of the following mitzvot:
Make meals for Cooking 4 Comfort

L'Shanah Tovah, may the new year be a meaningful one for you."

4. On the Spiritual page, when I click other and move the cursor into the box it displays all of the other items that have been there.  I don't know if it's just the ones on my computer.

Her Item #4 is her auto-fill issue
Removed extraneous files